Fix multiple invocations of $watch
Apparently, when you $watch on a $rootScope, this watches didn't get cleared when you switch your application and tend to pile up indefinitely creating a very nasty leak resulting in a flood of requests to the backend. On the other hand, local $scope, while providing an access to the same root scope variable, gets destructed and clears it's watches on an application switch.

Same with st2Api scopes. Apparently, 'list()' watches also have a tendency to pile up and cause a wide variety of problems down the road. I intentionaly left $scope._api in a local scope as I'm not sure what exactly am I going to do with the knowledge I've just acquired and it's a bit too soon to try and optimize this things.