Remove st2Loader for good
It leaks DOM references and I don't see a simple way to fix it. More than that, it was initially hacked very quickly, doesn't fit the needs for the rest of the app and generally have very little use since we've managed to fix API performance to the level we don't need additional indication to keep user involved.