WebUI does not auto refresh with clean browser
With a clean Browser cache/cookies, when the UI was launched, the
call to /stream used to fail due to which the UI did not refresh.
This PR fixes the issue by sending the auth-token in the /stream
request. This PR also updates the version of Eventsource library to
the latest version (1.0.7) which allows headers to be set in the
Eventsource request.